style: Bank관련 클래스, 구조체들 Bank.swift로 이동
